stun::agent

Struct Agent

pub struct Agent { /* private fields */ }

Agent is low-level abstraction over transaction list that handles concurrency (all calls are goroutine-safe) and time outs (via Collect call).

Implementations

impl Agent

pub fn new(handler: Handler) -> Self

new initializes and returns new Agent with provided handler.

pub fn stop_with_error( &mut self, id: TransactionId, error: Error, ) -> Result<(), Error>

stop_with_error removes transaction from list and calls handler with provided error. Can return ErrTransactionNotExists and ErrAgentClosed.

pub fn process(&mut self, message: Message) -> Result<(), Error>

process incoming message, synchronously passing it to handler.

pub fn close(&mut self) -> Result<(), Error>

close terminates all transactions with ErrAgentClosed and renders Agent to closed state.

pub fn start( &mut self, id: TransactionId, deadline: Instant, ) -> Result<(), Error>

start registers transaction with provided id and deadline. Could return ErrAgentClosed, ErrTransactionExists.

Agent handler is guaranteed to be eventually called.

pub fn stop(&mut self, id: TransactionId) -> Result<(), Error>

stop stops transaction by id with ErrTransactionStopped, blocking until handler returns.

pub fn collect(&mut self, deadline: Instant) -> Result<(), Error>

collect terminates all transactions that have deadline before provided time, blocking until all handlers will process ErrTransactionTimeOut. Will return ErrAgentClosed if agent is already closed.

It is safe to call Collect concurrently but makes no sense.

pub fn set_handler(&mut self, h: Handler) -> Result<(), Error>

set_handler sets agent handler to h.